python/requirements.txt
beautifulsoup4==4.13.3
psutil==6.1.1
python-dotenv==1.0.1
Requests==2.32.3
